% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessdb.R
\name{datasets}
\alias{datasets}
\title{Retrieve Datasets Table}
\usage{
datasets(ordered = TRUE)
}
\arguments{
\item{ordered}{logical. \code{TRUE} orders the result in the order data was entered into the database, while \code{FALSE} returns the result in a random order, to the benefit of faster query execution.}
}
\description{
This function pulls a table called 'DATASET' from the database, and returns it as a \emph{data.table} in R.
The 'DATASET' table gives information about the different datasets read into the database from various sources.
}
\examples{
\donttest{
datasets()
}
}
\seealso{
\code{\link{datasources}}, \code{\link{series}}, \code{\link{ugatsdb}}
}
