% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessdb.R
\name{datasources}
\alias{datasources}
\title{Retrieve Data Sources Table}
\usage{
datasources(ordered = TRUE)
}
\arguments{
\item{ordered}{logical. \code{TRUE} orders the result in the order data was entered into the database, while \code{FALSE} returns the result in a random order, to the benefit of faster query execution.}
}
\description{
This function pulls a table called 'DATASOURCE' from the database, and returns it as a \emph{data.table} in R.
The 'DATASOURCE' table gives information about the various sources / providers of data in this database.
}
\examples{
\donttest{
datasources()
}
}
\seealso{
\code{\link{datasets}}, \code{\link{ugatsdb}}
}
