% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessdb.R
\name{ugatsdb_reconnect}
\alias{ugatsdb_reconnect}
\title{Reconnect to Database}
\usage{
ugatsdb_reconnect()
}
\description{
This function terminates an existing connection to the database server and attempts to reconnect to it.
It is a useful function to include at the beginning of codes that should work regardless of whether the package was loaded recently or the internet connection is stable.
}
\examples{
\donttest{
ugatsdb_reconnect()
}
}
\seealso{
\code{\link{ugatsdb}}
}
