% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessdb.R
\name{datasources}
\alias{datasources}
\title{Retrieve Data Sources Table}
\usage{
datasources(ordered = TRUE)
}
\arguments{
\item{ordered}{logical. \code{TRUE} orders the result in the order data was entered into the database, while \code{FALSE} returns the result in a random order, to the benefit of faster query execution.}
}
\value{
A \code{\link[data.table]{data.table}} with information about the sources of data in the database.
}
\description{
This function pulls and returns a table called 'DATASOURCE' from the database.
}
\details{
The 'DATASOURCE' table gives information about the various sources / providers of data in this database, including the source website, the number of datasets available from the source, a description of the source and the way data is accessed from the source.
}
\examples{
\donttest{
datasources()
}
}
\seealso{
\code{\link{datasets}}, \code{\link{ugatsdb}}
}
