% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessdb.R
\name{ugatsdb_reconnect}
\alias{ugatsdb_reconnect}
\title{Reconnect to Database}
\usage{
ugatsdb_reconnect()
}
\description{
This function terminates an existing connection to the database server and attempts to reconnect to it.
It is now somewhat redundant by the safe query mechanism introduced in v0.2.1 of the package, where each query is evaluated inside
\code{\link{tryCatch}} and the database connection is renewed if the query fails. This function can still be used to manually
renew the database connection.
}
\examples{
\donttest{
ugatsdb_reconnect()
}
}
\seealso{
\code{\link{ugatsdb}}
}
